I just committed the final round of updates to ajdoc needed for 1.2.  In a nutshell it's moved forward a ways from the previous ajdoc: - AJDT integration works - generates documentation for all the AspectJ modules - works on 1.3 and 1.4 - comments and Javadoc tags are properly preserved and resolved
Note that the hope is still for "declare" documentation to come in the form of a contribution.  

Getting it to work for our tree involved a bunch of bug fixes to deal with things like anonymous types.  To run from our tree in bootstrap mode use something like the attached batch script.  

I had to roll back some of Andy's fix to broken links: getRelativeComponent didn't work properly for internal anchor links (..html#<mumble>).  Andy, could you check this against the current thing against whatever was breaking for you, and ideally add that code patter to CoverageTestCase?

I added -XjavadocsInModel to make the addition of Javadoc strings to the ASM only happen when request it (prevent bloat of the model with strings.  The running time of ajdoc is now dominated by the compilation time.