Ensure we can disassemble invokedynamic
If there are problems at weave time the weaver may attempt to
disassemble some code to produce a nice error message. Until this
change that disassembly code did not understand invokedynamic.
This would make it fail to disassemble and instead of seeing the
real problem you see a disassembly problem. With this fix we
should now see the underlying problem in the issue.

Issue: #525541