On JDK11 call the defineClass variant with a ProtectionDomain param
This alters the defineClass used when loadtime weaving and defining
'extra' classes (e.g. closures). This change causes us to call
the defineClass() method variant on ClassLoader that takes a
protection domain, so it should share the same protection domain as
the class for which the extra class was generated.

Issue: https://bugs.eclipse.org/bugs/show_bug.cgi?id=542421