Remove 'imagesdir' from Asciidoctor Maven config
User relative paths for images instead.

In one case, we need to copy images from a subdirectory one level up,
because otherwise the images are only found in the included document,
but not in the single-page one including it from the subdirectory. In
all other cases, included ADOCs are located in the same directory as the
including document.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>