Extend documentation for Antrun execution to fix javadocs
Also update the internal review ID of the OpenJDK bug I created, because
previously I had created a feature request, which has remained
unanswered. But actually, I think it is rather a regression bug, having
found out that before JDK 16 the behaviour was exactly how the feature
request describes it.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>