Remove lib/jdtcore-aj
If any of the old Ant builds, e.g. tests/profiling/build.xml, which
have never been mavenised, need JDT Core, they should be converted to
Maven builds and refer to it as a regular dependency. As is, the Ant
builds would not run anyway, because other dependency locations have
changed as well.

Signed-off-by: Alexander Kriegisch <Alexander@Kriegisch.name>